% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illdth-cif-data-analysis.r
\name{cifcmp.full}
\alias{cifcmp.full}
\title{Apply three methods (see note below) for estimating the cumulative incidence function for disease in an illness-death model.}
\usage{
cifcmp.full(
  age_recr,
  age_diag,
  age_death,
  age_end_fu,
  status_end,
  tgrd,
  tgrd.cb,
  covpr,
  nresam,
  dtgrd_width,
  bwselect,
  bwvec,
  nfld,
  redo_bw,
  verbosity
)
}
\arguments{
\item{age_recr}{age at recruitment}

\item{age_diag}{age at diagnosis}

\item{age_death}{age at death}

\item{age_end_fu}{age at end of follow-up (death or censoring)}

\item{status_end}{status at end of follow-up}

\item{tgrd}{grid of ages at which the CIF will be computed}

\item{tgrd.cb}{grid of ages over which the simultaneous confidence band will be computed (subset of tgrd)}

\item{covpr}{desired confidence interval coverage probability}

\item{nresam}{number of bootstrap replications for confidence band}

\item{dtgrd_width}{grid width for age at diagnosis in conditional survival calculation}

\item{bwselect}{TRUE/FALSE flag for whether to carry out bandwidth selection}

\item{bwvec}{vector of candidate bandwidths}

\item{nfld}{number of folds in crossvalidation procedure for selecting the bandwidth}

\item{redo_bw}{TRUE/FALSE flag for redoing bandwidth selection in bootstrap}

\item{verbosity}{code for verbosity level (0=none, 1=intermediate, 2=full)}
}
\value{
a list containing the estimates, pointwise confidence limits, and simultaneous confidence band limits
for all three methods and the bandwidth parameter for the Zucker-Gorfine method
}
\description{
Apply three methods (see note below) for estimating the cumulative incidence function for disease in an illness-death model.
}
\note{
Methods provided: \cr
1. The method of Aalen and Johansen (1978) \cr
2. The method of Gorfine, Zucker, and Shoham (2025) \cr
3. The method of Zucker and Gorfine (2026)

Method #3 includes all disease cases.

status_end = status at end of follow-up \cr
0 = alive without disease (i.e. censored) \cr
1 = died without disease \cr
2 = alive with disease \cr
3 = died with disease \cr

When no bandwidth selection is done, the bandwidth used is bwvec[1].
}
\references{
Aalen, O. O. and Johansen, S. (1978). An empirical transition matrix for non-homogeneous Markov chains
based on censored observations. Scandinavian Journal of Statistics, 5:141–150.

Gorfine, M., Zucker, D. M., and Shoham, S. (2025). Cumulative incidence function estimation using populationbased
biobank data. Biometrics, 81(3):ujaf049.

Zucker, D. M. and Gorfine, M. Cumulative incidence function
estimation in the illness-death model using all disease cases. arXiv.
}
